V 0.5.3

+++ TerraIncognita2/Assets/Source/Scripts/Base/CollectorsSpawner.cs
@@ -19,5 +19,5 @@ public class CollectorsSpawner : MonoBehaviour
         return collector;
     }
 
-    public float GetCollectorPrice() => _collectorBlueprint.Price;    
+    public int GetCollectorPrice() => _collectorBlueprint.Price;    
 }
